Implemented case MediaFileUpdated in CacheDatabaseManager, cleanup

+++ HiP-DataStore.Model/Rest/SliderPageImageArgs.cs
@@ -35,7 +35,7 @@ namespace PaderbornUniversity.SILab.Hip.DataStore.Model.Rest
         public SliderPageImageResult(SliderPageImage img)
         {
             Date = img.Date;
-            Image = (int)img.Image;
+            Image = img.Image;
         }
     }
 }
